MT_File_Downloader/requirements.txt
alabaster==0.7.10
astroid==1.6.0
attrs==17.4.0
Babel==2.5.1
beautifulsoup4==4.6.0
bleach==2.1.2
certifi==2017.11.5
cffi==1.11.4
chardet==3.0.4
cloudpickle==0.5.2
cmake==0.9.0
colorama==0.3.9
comtypes==1.1.4
cycler==0.10.0
Cython==0.27.3
decorator==4.1.2
docutils==0.14
entrypoints==0.2.3
enum34==1.1.6
flake8==3.5.0
html5lib==1.0.1
idna==2.6
imagesize==0.7.1
ipykernel==4.7.0
ipython==6.2.1
ipython-genutils==0.2.0
ipywidgets==7.1.0
isort==4.2.15
jedi==0.11.1
Jinja2==2.10
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.2.0
jupyter-console==5.2.0
jupyter-core==4.4.0
jupyterlab==0.30.6
jupyterlab-launcher==0.6.0
Keras==2.1.2
lazy-object-proxy==1.3.1
lightgbm==2.0.12
lxml==4.2.1
Markdown==2.6.10
MarkupSafe==1.0
matplotlib==2.1.1
mccabe==0.6.1
mistune==0.8.3
MulticoreTSNE==0.1
nbconvert==5.3.1
nbformat==4.4.0
notebook==5.2.2
numerapi==0.0.0
numpy==1.14.0rc1+mkl
numpydoc==0.7.0
pandas==0.21.1
pandocfilters==1.4.2
parso==0.1.1
patsy==0.5.0
pickleshare==0.7.4
Pillow==5.0.0
pluggy==0.6.0
prompt-toolkit==1.0.15
protobuf==3.5.1
psutil==5.4.3
py==1.5.2
pycodestyle==2.3.1
pycparser==2.18
pyflakes==1.6.0
Pygments==2.2.0
pylint==1.8.1
pyparsing==2.2.0
PyQt5==5.9.2
pystan==2.17.1.0
pytest==3.3.1
python-dateutil==2.6.1
python-docx==0.8.6
pytz==2017.3
PyYAML==3.12
pyzmq==16.0.3
QtAwesome==0.4.4
qtconsole==4.3.1
QtPy==1.3.1
requests==2.18.4
rope==0.10.7
scikit-learn==0.19.1
scipy==1.0.0
seaborn==0.8.1
selenium==3.12.0
simplegeneric==0.8.1
sip==4.19.6
six==1.11.0
sklearn==0.0
snowballstemmer==1.2.1
Sphinx==1.6.5
sphinxcontrib-websupport==1.0.1
statsmodels==0.8.0
tensorflow==1.4.0
tensorflow-gpu==0.12.1
tensorflow-tensorboard==0.4.0rc3
testpath==0.3.1
tornado==4.5.2
traitlets==4.3.2
urllib3==1.22
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.13
widgetsnbextension==3.1.0
win-unicode-console==0.5
wordcloud==1.3.1
wrapt==1.10.11
xgboost==0.7
